% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermixmod.R
\name{clustermixmod}
\alias{clustermixmod}
\title{Mixture models for clustering}
\usage{
clustermixmod(
  x,
  K,
  itermax = 100,
  init = c("random", "kmeans"),
  datatype = c("continuous", "categorical", "mix")
)
}
\arguments{
\item{x}{a data frame containing quantitative,qualitative or heterogeneous data. Rows
correspond to observations and columns correspond to variables.}

\item{K}{numeric listing the number of clusters.}

\item{itermax}{numeric. The max of iteration.}

\item{init}{character. Type of initialization is "random" or "kmeans".}

\item{datatype}{character. Type of data is "continuous", "categorical" or "mix".}
}
\value{
prop: posteriori probabilities for each individual to belong to each cluster.

mu, sigma: mean and matrix correlation of each cluster for data "continuous".

loglik: log-likelihood of model.

iterfin: the last iteration.

K:  numeric listing the number of clusters.

z: cluster of each individual.

BIC: criterion BIC of model.

ICL: criterion ICL of model.
}
\description{
This function fitting a mixture model clustering of multivariate gaussian or multinomial components or the both to a given dataset.
}
