% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical.R
\name{plot_categorical}
\alias{plot_categorical}
\title{plot result of Latent Class model}
\usage{
plot_categorical(x, model)
}
\arguments{
\item{x}{a data frame containing quantitative,qualitative or heterogeneous data. Rows
correspond to observations and columns correspond to variables.}

\item{model}{model of Latent Class Model from function clustermixmod()}
}
\value{
a graph represents result of model
}
\description{
This function plot result of Latent Class model (type of data is categorical).
}
