% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuous.R
\name{plot_continuous}
\alias{plot_continuous}
\title{plot result of Gaussian mixture model}
\usage{
plot_continuous(x, model)
}
\arguments{
\item{x}{a data frame containing quantitative data. Rows
correspond to observations and columns correspond to variables.}

\item{model}{model of Gaussian mixture model from function clustermixmod()}
}
\value{
a graph represents result of model
}
\description{
This function plot result of Gaussian mixture model (type of data is continuous).
}
