% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mix.R
\name{plot_mix}
\alias{plot_mix}
\title{plot result of miture model}
\usage{
plot_mix(x, model)
}
\arguments{
\item{x}{a data frame containing heterogeneous data. Rows
correspond to observations and columns correspond to variables.}

\item{model}{mixture model from function clustermixmod()}
}
\value{
a graph represents result of model
}
\description{
This function plot result of mixturemodel (type of data is both of continuous and categorical).
}
